fix: 修复关闭快速加载可能因为找不到idx文件而崩溃, ui细节调整 feat: 支持检测缓存完整和idx文件是否存在

+++ app/src/main/java/io/github/kineks/neteaseviewer/data/local/CacheFileInfo.kt
@@ -8,7 +8,7 @@ data class CacheFileInfo(
     @SerializedName("filemd5")
     val fileMD5: String,
     @SerializedName("filesize")
-    val fileSize: Int,
+    val fileSize: Long,
     val md5: String,
     @SerializedName("musicId")
     val id: Int,
